Add Net::AMQP::RabbitMQ to the list of dependencies
Net::RabbitFoot is used in RabbitMQ's code examples but it downright
refuses to install on OS X due to Coro dependency's compile bugs.
Net::AMQP::RabbitMQ seems decent, uses native library for communicating
with RabbitMQ, doesn't have that many failed unit tests on CPAN and
would be easy to replace with an alternative.